chore(cli): retire dead generic `gaia update merge` command
The generic three-way file walk under `gaia update merge` has no callers.
The /update-gaia skill invoked it until the Step 7/9 rewrite (#231), which
moved to hand-walking the decision table and field-merging package.json
(7a) and pnpm-workspace.yaml (7b) by name. The generic command has been
orphaned since, with only janitorial touches (deps, react-doctor, em-dash
sweeps), no behavioral use.

Its two helpers (three-way.ts, util/manifest.ts) had no external callers,
so retiring the command retires them too. The field-aware merge-workspace
oracle (the live Step 7b CLI) is untouched.

- Delete merge.ts, merge.test.ts, util/three-way.ts, util/manifest.ts
- Unwire `merge` from the update router and both binaries' help text
- Add a router guard test: `update merge` now resolves unknown_subcommand
- Delete the stale wiki/concepts/Update Merge.md (claimed the skill invoked
  the command and that it could be run manually, both false and a footgun:
  a manual run whole-file-merges the two field-aware files)
- Drop its wiki/index.md catalog entry and the dangling pnpm.md wikilink
- Regenerate manifest (drops the deleted wiki page; --check clean)
- Rebuild both binaries
