+++ .gaia/cli/src/index.maintainer.ts
@@ -40,7 +40,7 @@ Maintainer-only binary. Adopters use 'gaia' (no release namespace).
   wiki state|commit-classify|state-init|state-bump|log-prepend|page-index|orphans|near-collisions|dead-paths|sync land
   fitness render-card [--cols N]
   automation read-config|read-state|init-state|bump-state|cron-decide|record-run|record-overage|clear-overage
-  update merge --baseline <dir> --latest <dir> --manifest <path>
+  update merge-workspace --baseline <file> --latest <file> --current <file>
   update-deps run --emit-updates <path>
   release preflight|bump|changelog|scrub-wiki|manifest|scrub|runtime-deps|commit-and-tag
   init strip-branding|configure-i18n|rename|wire-statusline|finalize|resume
